+++ data-agent-management/src/main/java/com/alibaba/cloud/ai/dataagent/controller/AgentDatasourceController.java
@@ -47,7 +47,7 @@ public class AgentDatasourceController {
 	 * Information Source" function on the frontend
 	 */
 	@PostMapping("/init")
-	public ApiResponse<?> initSchema(@PathVariable Long agentId) {
+	public ApiResponse<?> initSchema(@PathVariable Integer agentId) {
 		// 防止前端恶意请求，dto数据应该在后端获取
 		try {
 			AgentDatasource agentDatasource = agentDatasourceService.getCurrentAgentDatasource(agentId);
@@ -86,7 +86,7 @@ public class AgentDatasourceController {
 
 	/** Get list of data sources configured for agent */
 	@GetMapping
-	public ApiResponse<List<AgentDatasource>> getAgentDatasource(@PathVariable Long agentId) {
+	public ApiResponse<List<AgentDatasource>> getAgentDatasource(@PathVariable Integer agentId) {
 		try {
 			log.info("Getting datasources for agent: {}", agentId);
 			List<AgentDatasource> datasources = agentDatasourceService.getAgentDatasource(agentId);
@@ -100,7 +100,7 @@ public class AgentDatasourceController {
 	}
 
 	@GetMapping("/active")
-	public ApiResponse<AgentDatasource> getActiveAgentDatasource(@PathVariable Long agentId) {
+	public ApiResponse<AgentDatasource> getActiveAgentDatasource(@PathVariable Integer agentId) {
 		try {
 			log.info("Getting active datasource for agent: {}", agentId);
 			AgentDatasource datasource = agentDatasourceService.getCurrentAgentDatasource(agentId);
@@ -114,7 +114,7 @@ public class AgentDatasourceController {
 
 	/** Add data source for agent */
 	@PostMapping("/{datasourceId}")
-	public ApiResponse<AgentDatasource> addDatasourceToAgent(@PathVariable Long agentId,
+	public ApiResponse<AgentDatasource> addDatasourceToAgent(@PathVariable Integer agentId,
 			@PathVariable Integer datasourceId) {
 		try {
 			if (datasourceId == null) {
@@ -131,7 +131,7 @@ public class AgentDatasourceController {
 
 	// 更新选择的数据表
 	@PostMapping("/tables")
-	public ApiResponse<?> updateDatasourceTables(@PathVariable Long agentId,
+	public ApiResponse<?> updateDatasourceTables(@PathVariable Integer agentId,
 			@RequestBody @Validated UpdateDatasourceTablesDTO dto) {
 		try {
 			dto.setTables(Optional.ofNullable(dto.getTables()).orElse(List.of()));
@@ -146,7 +146,7 @@ public class AgentDatasourceController {
 
 	/** Remove data source association from agent */
 	@DeleteMapping("/{datasourceId}")
-	public ApiResponse<?> removeDatasourceFromAgent(@PathVariable Long agentId, @PathVariable Integer datasourceId) {
+	public ApiResponse<?> removeDatasourceFromAgent(@PathVariable Integer agentId, @PathVariable Integer datasourceId) {
 		try {
 			agentDatasourceService.removeDatasourceFromAgent(agentId, datasourceId);
 			return ApiResponse.success("数据源已移除");
@@ -158,7 +158,7 @@ public class AgentDatasourceController {
 
 	/** 启用/禁用智能体的数据源 */
 	@PutMapping("/toggle")
-	public ApiResponse<AgentDatasource> toggleDatasourceForAgent(@PathVariable Long agentId,
+	public ApiResponse<AgentDatasource> toggleDatasourceForAgent(@PathVariable Integer agentId,
 			@RequestBody ToggleDatasourceDTO dto) {
 		try {
 			Boolean isActive = dto.getIsActive();
