+++ data-agent-management/src/main/java/com/alibaba/cloud/ai/dataagent/entity/SemanticModel.java
@@ -34,12 +34,12 @@ public class SemanticModel {
 	/**
 	 * 主键ID
 	 */
-	private Long id;
+	private Integer id;
 
 	/**
 	 * 关联的智能体ID
 	 */
-	private Long agentId;
+	private Integer agentId;
 
 	/**
 	 * 关联的数据源ID
